Add python3 port of deps/resolve-deps.
Adjust the test harness to take an initial resolve-deps program path to
be able to test either the ClojureScript or python3 version.

Also, fix a minor bug discovered in the Clojure implementation: removing
the start node in the wrong place. This should be done by the caller and
not in the resolve-dep-order function so that the function is more
generally usable.

The python3 version is also substantially faster.

Note: this port was done by leveraging ChatGPT 4.0. It was able to port
98% of the code in seconds. However, it also generated half a dozen
logic bugs that I had to track down and fix. Most were around handling
alternate dependencies lists: either trying to use them for lookups in
the visited set (which works in Clojure and always returns false) or
failing to iterate those alternate dependencies in the right places.
Overall, using ChatGPT for this task probably saved 3-4 hours of time.