Add support for loading JSON dep files.
If the component of the path specifier refers to a file rather than
a directory, then treat it as a JSON file containing a dependency
specification. The JSON file uses the full dependency map/graph format
rather than the user dep string format of deps file within dep
directories. For example, the following JSON file specifies that "a"
depends on both "b" and "c", that "c" depends on either "d" of "e", and
that "f" must come after "c" if it required:

```
{
    "a": ["b", "c"],
    "c": [{"or": ["d", "e"]}],
    "f": [{:after "c"}]
}
```

If a path component is "-", then dependencies will be read as a JSON
string from stdin.

Also add equivalent JSON files for each dependency directory under
`./tests/`.