Extract/expose graph conversion/normalization functions.
Extract the full-to-alt-graph and alt-to-kahn-graph functions so that
they can be called by users of the library functions since they are
needed to fully utilitize the functions.

Change the order of arguments to the python alt_set_covers function so
that the pending argument is first. This allows direct calling of
alt_set_covers with two position arguments (instead of requiring the
pending argument to be passed by name).